requirements.txt
aiogram
aiohttp
aioredis
gino
python-dotenv